Added a Pod class to prepare Game populated by main menu state; called GameMode for the PongFramework.
Added a helper class for MenuNav to move menu element.

Added MenuCursor to hold the sprite for a cursor and it's navigation utilizing MenuNav.

Added a paginated menu to roll through multiple pages of the menu. This has been placed but not finalized within main menu state for the PongFramework.